requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
contourpy==1.1.0
cycler==0.11.0
Django==4.1
django-cors-headers==4.2.0
djangorestframework==3.14.0
fonttools==4.42.1
importlib-resources==6.0.1
kiwisolver==1.4.5
matplotlib==3.7.2
mpmath==1.3.0
mysqlclient==2.2.0
numpy==1.24.4
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
sympy==1.12
typing_extensions==4.7.1
tzdata==2023.3
zipp==3.16.2
scipy
gTTS==2.5.0
PyPDF4==1.27.0